fix: moved @core providers to ui-sdk/core
