--- apps/gauzy/src/app/@core/index.ts
@@ -1,2 +1,0 @@
-export * from './auth';
-export * from './core.module';
