+++ apps/gauzy/src/app/@theme/components/gauzy-logo/gauzy-logo.component.html
@@ -1,12 +1,9 @@
 <nb-accordion *ngIf="isAccordion" class="accordion workspace">
-	<nb-accordion-item
-		[collapsed]="controlled"
-		(collapsedChange)="onCollapse($event)"
-	>
+	<nb-accordion-item [collapsed]="controlled" (collapsedChange)="onCollapse($event)">
 		<nb-accordion-item-header class="principal">
 			<ng-container *ngTemplateOutlet="logo"></ng-container>
 			<div class="organization" *ngIf="!isCollapse">
-				<img [src]="organization?.imageUrl" [alt]="organization?.name"/>
+				<img [src]="organization?.imageUrl" [alt]="organization?.name" />
 				<div class="description text">
 					<div [nbTooltip]="organization?.name">{{ organization?.name }}</div>
 					<div [nbTooltip]="organization?.website">{{ organization?.website }}</div>
@@ -30,17 +27,20 @@
 <ng-template #logo>
 	<div
 		*ngIf="isCollapse"
-		[ngClass]="{'white-svg': isSVG && (theme === 'dark' || theme === 'cosmic' || theme === 'gauzy-dark' || theme === 'material-dark')}"
+		[ngClass]="{
+			'white-svg':
+				isSVG && (theme === 'dark' || theme === 'cosmic' || theme === 'gauzy-dark' || theme === 'material-dark')
+		}"
 		class="logo"
 		(click)="navigateHome()"
 	>
 		<ng-container *ngIf="isSVG; else image">
 			<object [data]="logoUrl" type="image/svg+xml">
-				<img src='assets/images/logos/logo_Gauzy.png'/>
+				<img src="/assets/images/logos/logo_Gauzy.png" />
 			</object>
 		</ng-container>
 		<ng-template #image>
-			<img [src]="logoUrl || 'assets/images/logos/logo_Gauzy.png'"/>
+			<img [src]="logoUrl || '/assets/images/logos/logo_Gauzy.png'" />
 		</ng-template>
 	</div>
 </ng-template>
