+++ apps/gauzy/src/app/@theme/components/gauzy-logo/gauzy-logo.component.ts
@@ -13,8 +13,8 @@ import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { tap, debounceTime, filter } from 'rxjs/operators';
 import { NbThemeService } from '@nebular/theme';
 import { untilDestroyed, UntilDestroy } from '@ngneat/until-destroy';
-import { Store, distinctUntilChange } from '@gauzy/ui-sdk/common';
 import { FeatureEnum, IOrganization, PermissionsEnum } from '@gauzy/contracts';
+import { Store, distinctUntilChange } from '@gauzy/ui-sdk/common';
 import { Environment, GAUZY_ENV } from '@gauzy/ui-config';
 
 @UntilDestroy({ checkProperties: true })
@@ -207,8 +207,8 @@ export class GauzyLogoComponent implements AfterViewInit, OnInit, OnDestroy {
 		this.store.selectedOrganization$
 			.pipe(
 				debounceTime(100),
-				distinctUntilChange(),
 				filter((organization: IOrganization) => !!organization),
+				distinctUntilChange(),
 				tap((organization: IOrganization) => (this.organization = organization)),
 				untilDestroyed(this)
 			)
