+++ apps/gauzy/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { ExtraOptions, RouterModule, Routes } from '@angular/router';
 import { NgModule } from '@angular/core';
-import { AuthGuard } from './@core/auth/auth.guard';
+import { AuthGuard } from '@gauzy/ui-sdk/core';
 import { SignInSuccessComponent } from './auth/signin-success/sign-in-success.component';
 import { SignInSuccessModule } from './auth/signin-success/signin-success-login-google.module';
 import { AppModuleGuard } from './app.module.guards';
@@ -54,4 +54,4 @@ const config: ExtraOptions = {
 	],
 	exports: [RouterModule]
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
