+++ apps/gauzy/src/app/pages/pages.component.ts
@@ -8,14 +8,17 @@ import { NgxPermissionsService } from 'ngx-permissions';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { chain } from 'underscore';
 import { TranslationBaseComponent } from '@gauzy/ui-sdk/shared';
-import { AuthStrategy, NavMenuBuilderService, NavMenuSectionItem, UsersService } from '@gauzy/ui-sdk/core';
-import { FeatureEnum, IOrganization, IRolePermission, IUser, IntegrationEnum, PermissionsEnum } from '@gauzy/contracts';
-import { Store, distinctUntilChange, isNotEmpty } from '@gauzy/ui-sdk/common';
 import {
+	AuthStrategy,
 	IJobMatchingEntity,
 	IntegrationEntitySettingServiceStoreService,
-	IntegrationsService
+	IntegrationsService,
+	NavMenuBuilderService,
+	NavMenuSectionItem,
+	UsersService
 } from '@gauzy/ui-sdk/core';
+import { FeatureEnum, IOrganization, IRolePermission, IUser, IntegrationEnum, PermissionsEnum } from '@gauzy/contracts';
+import { Store, distinctUntilChange, isNotEmpty } from '@gauzy/ui-sdk/common';
 import { ReportService } from './reports/all-report/report.service';
 
 @UntilDestroy({ checkProperties: true })
