+++ apps/gauzy/src/app/pages/pages.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { NbMenuModule, NbToastrModule, NbSpinnerModule, NbIconModule } from '@nebular/theme';
 import { FeatureToggleModule as NgxFeatureToggleModule } from 'ngx-feature-toggle';
-import { AuthService, RoleGuard } from '@gauzy/ui-sdk/core';
+import { AuthService, CommonNavModule, RoleGuard } from '@gauzy/ui-sdk/core';
 import { I18nTranslateModule } from '@gauzy/ui-sdk/i18n';
 import { ThemeModule } from '../@theme/theme.module';
 import { PagesComponent } from './pages.component';
 import { PagesRoutingModule } from './pages-routing.module';
 import { MiscellaneousModule } from './miscellaneous/miscellaneous.module';
-import { CommonNavModule } from '../@core/components/common-nav.module';
 
 @NgModule({
 	imports: [
