+++ packages/ui-sdk/package.json
@@ -9,8 +9,6 @@
 	},
 	"license": "AGPL-3.0",
 	"private": true,
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
 	"files": [
 		"dist"
 	],
@@ -32,7 +30,12 @@
 		"@angular/core": "^16.2.12"
 	},
 	"dependencies": {
+		"@angular/material": "^16.2.12",
+		"@angular/router": "^16.2.12",
 		"@gauzy/ui-config": "^0.1.0",
+		"@nebular/auth": "^12.0.0",
+		"@nebular/security": "^12.0.0",
+		"@nebular/theme": "^12.0.0",
 		"@ngneat/until-destroy": "^9.2.0",
 		"@ngx-translate/core": "^14.0.0",
 		"@ngx-translate/http-loader": "^7.0.0",
