+++ packages/ui-sdk/src/lib/core/src/auth/no-auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Store } from '@gauzy/ui-sdk/common';
-import { AuthService } from '@gauzy/ui-sdk/core';
+import { AuthService } from '../services';
 /**
  * Use for routes which only need to be displayed if user is NOT logged in
  */
