+++ packages/ui-sdk/src/lib/core/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './common-nav.module';
+export * from './base-nav-menu/base-nav-menu.component';
+export * from './main-nav-menu/main-nav-menu.component';
+export * from './sidebar-menu/sidebar-menu.component';
