+++ packages/ui-sdk/src/lib/core/src/components/main-nav-menu/main-nav-menu.component.html
