+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/index.ts
@@ -1,0 +1,4 @@
+export * from './sidebar-menu.component';
+export * from './tooltip.directive';
+export * from './menu-items/concrete/menu-item/menu-item.component';
+export * from './menu-items/concrete/children-menu-item/children-menu-item.component';
