+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/menu-items/concrete/children-menu-item/children-menu-item.component.html
