+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/menu-items/concrete/menu-item/menu-item.component.scss
