+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/menu-items/concrete/menu-item/menu-item.component.ts
@@ -4,8 +4,8 @@ import { Router } from '@angular/router';
 import { tap } from 'rxjs/operators';
 import { NbSidebarService } from '@nebular/theme';
 import { IUser } from '@gauzy/contracts';
-import { JitsuAnalyticsEvents, JitsuAnalyticsEventsEnum, JitsuService } from '@gauzy/ui-sdk/core';
 import { Store } from '@gauzy/ui-sdk/common';
+import { JitsuAnalyticsEvents, JitsuAnalyticsEventsEnum, JitsuService } from '../../../../../services';
 import { IMenuItem } from '../../interface/menu-item.interface';
 
 @Component({
