+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/menu-items/interface/menu-item.interface.ts
