+++ packages/ui-sdk/src/lib/core/src/components/sidebar-menu/sidebar-menu.component.scss
