+++ packages/ui-sdk/src/lib/core/src/services/nav-builder/index.ts
@@ -1,2 +1,3 @@
 export * from './nav-builder-types';
 export * from './nav-menu-builder.service';
+export * from './sidebar-menu.service';
