+++ packages/ui-sdk/src/lib/core/src/services/nav-builder/sidebar-menu.service.ts
