+++ packages/ui-sdk/src/lib/ng-package.json
@@ -1,6 +1,6 @@
 {
 	"$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-	"dest": "../../../../dist/packages/ui-sdk",
+	"dest": "../../dist",
 	"allowedNonPeerDependencies": ["."],
 	"assets": ["./i18n/assets/i18n/*.*", "./static/styles/**/*.scss"],
 	"lib": {
